--- src/main/java/com/dementor/domain/categories/entity/Job.java
@@ -1,38 +1,0 @@
-package com.dementor.domain.categories.entity;
-
-import java.util.List;
-
-import com.dementor.domain.mentor.entity.Mentor;
-
-import jakarta.persistence.CascadeType;
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.OneToMany;
-import jakarta.persistence.Table;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Entity
-@Table(name = "job")
-@Getter
-@NoArgsConstructor
-@AllArgsConstructor
-@Builder
-public class Job {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @Column(nullable = false)
-    private String name;
-
-    // Mentor 엔티티와의 관계 (일대다)
-    @OneToMany(mappedBy = "job", cascade = CascadeType.ALL)
-    private List<Mentor> mentors;
-}
-
