+++ src/main/java/com/dementor/domain/mentor/entity/Mentor.java
@@ -1,49 +1,31 @@
 package com.dementor.domain.mentor.entity;
 
-import java.util.List;
-
-import com.dementor.domain.categories.entity.Job;
+import com.dementor.domain.job.entity.Job;
 import com.dementor.domain.member.entity.Member;
 import com.dementor.domain.mentoringclass.entity.MentoringClass;
+import jakarta.persistence.*;
+import lombok.*;
 
-import jakarta.persistence.CascadeType;
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.EnumType;
-import jakarta.persistence.Enumerated;
-import jakarta.persistence.FetchType;
-import jakarta.persistence.Id;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.ManyToOne;
-import jakarta.persistence.MapsId;
-import jakarta.persistence.OneToMany;
-import jakarta.persistence.OneToOne;
-import jakarta.persistence.Table;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.Setter;
+import java.util.List;
 
 @Entity
 @Table(name = "mentor")
 @Getter
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
 @AllArgsConstructor
 @Builder
 public class Mentor {
 
     @Id
-    @Column(name = "member_id")
-    private Long id;
+    private Long memberId;
 
-    // Member 엔티티와의 관계 (일대일)
     @OneToOne(fetch = FetchType.LAZY)
     @MapsId // member_id를 PK이자 FK로 사용
-    @JoinColumn(name = "member_id")
+    @JoinColumn
     private Member member;
 
-    // categories 엔티티와의 관계 (다대일)
     @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "job_id", nullable = false)
+    @JoinColumn(nullable = false)
     private Job job;
 
     // PostAttachment 엔티티와의 관계 (일대일, 식별관계)
@@ -51,9 +33,12 @@ public class Mentor {
     //private PostAttachment attachment;
 
     // Mentoring 수업 엔티티와의 관계 (일대다)
-    @OneToMany(cascade = CascadeType.ALL)
-    @JoinColumn(name = "mentoringClass_id")
+    @OneToMany(cascade = CascadeType.ALL, orphanRemoval = true)
     private List<MentoringClass> mentorings;
+
+    // todo: 파일첨부 테이블과의 관계 (추후 구현)
+    //@Column(name = "attachment_id")
+    //private Long attachmentId;
 
     @Column(length = 10, nullable = false)
     private String name;
@@ -71,29 +56,15 @@ public class Mentor {
     private String introduction;
 
     @Setter
-    @Enumerated(EnumType.STRING)
     @Column(nullable = false)
-    private ApprovalStatus isApproved;
+    private Boolean isApproved;
 
     @Setter
-    @Enumerated(EnumType.STRING)
     @Column(nullable = false)
-    private ApprovalStatus isModified;
+    private Boolean isModified;
 
-    @Column(name = "best_for", length = 255)
+    @Column(length = 255)
     private String bestFor;
-
-    @Column(name = "attachment_id")
-    private Long attachmentId;
-
-	public Mentor() {
-	}
-
-	// ApprovalStatus Enum 정의
-    public enum ApprovalStatus {
-        Y, N
-    }
-
 
     // 멘토링 수업 추가 메서드
     public void addMentoringClass(MentoringClass mentoringClass) {
